refactoring and restructuring with focus on oidc-spring-support (massive commit in order to have it build properly) * first step on the road to remove the OIDC term and use more generic terms instead since the same code can be used to validate access tokens etc.
