+++ oidc-jersey-support/src/main/java/no/nav/security/token/support/core/jaxrs/servlet/JaxrsOIDCTokenValidationFilter.java
@@ -2,12 +2,12 @@ package no.nav.security.token.support.core.jaxrs.servlet;
 
 import no.nav.security.token.support.core.configuration.MultiIssuerConfiguration;
 import no.nav.security.token.support.core.filter.OIDCTokenValidationFilter;
-import no.nav.security.token.support.core.jaxrs.JaxrsJwtTokenContextHolder;
+import no.nav.security.token.support.core.jaxrs.JaxrsTokenContextHolder;
 import no.nav.security.token.support.core.validation.JwtTokenValidationHandler;
 
 public class JaxrsOIDCTokenValidationFilter extends OIDCTokenValidationFilter {
 
     public JaxrsOIDCTokenValidationFilter(MultiIssuerConfiguration oidcConfig) {
-        super(new JwtTokenValidationHandler(oidcConfig), JaxrsJwtTokenContextHolder.getHolder());
+        super(new JwtTokenValidationHandler(oidcConfig), JaxrsTokenContextHolder.getHolder());
     }
 }
