--- oidc-spring-support/src/main/java/no/nav/security/spring/oidc/api/EnableOIDCTokenValidation.java
@@ -1,28 +1,0 @@
-package no.nav.security.spring.oidc.api;
-/*
- * THIS CODE IS PROVIDED *AS IS* BASIS, WITHOUT WARRANTIES OR CONDITIONS
- * OF ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING WITHOUT LIMITATION
- * ANY IMPLIED WARRANTIES OR CONDITIONS OF TITLE, FITNESS FOR A
- * PARTICULAR PURPOSE, MERCHANTABILITY OR NON-INFRINGEMENT.
- */
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Inherited;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import org.springframework.context.annotation.Import;
-
-import no.nav.security.spring.oidc.EnableOIDCTokenValidationConfiguration;
-
-@Documented
-@Inherited
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-@Import({
-    EnableOIDCTokenValidationConfiguration.class
-})
-public @interface EnableOIDCTokenValidation {
-	String[] ignore() default {"org.springframework"};
-}
