--- oidc-spring-support/src/main/java/no/nav/security/spring/oidc/validation/interceptor/OIDCUnauthorizedException.java
@@ -1,14 +1,0 @@
-package no.nav.security.spring.oidc.validation.interceptor;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@SuppressWarnings("serial")
-@ResponseStatus(HttpStatus.UNAUTHORIZED)
-public class OIDCUnauthorizedException extends RuntimeException {
-	
-	public OIDCUnauthorizedException(String msg) {
-		super(msg);
-	}
-
-}
