+++ oidc-spring-support/src/main/java/no/nav/security/token/support/spring/SpringTokenContextHolder.java
@@ -1,0 +1,33 @@
+package no.nav.security.token.support.spring;
+
+import no.nav.security.token.support.core.context.TokenValidationContext;
+import org.springframework.web.context.request.RequestAttributes;
+import org.springframework.web.context.request.RequestContextHolder;
+
+import no.nav.security.token.support.core.context.TokenContextHolder;
+
+import java.util.Collections;
+
+public class SpringTokenContextHolder implements TokenContextHolder {
+
+    private static final String TOKEN_VALIDATION_CONTEXT_ATTRIBUTE = SpringTokenContextHolder.class.getName();
+
+	@Override
+	public TokenValidationContext getTokenValidationContext() {
+		TokenValidationContext tokenValidationContext = (TokenValidationContext)getRequestAttribute(TOKEN_VALIDATION_CONTEXT_ATTRIBUTE);
+		return tokenValidationContext != null ? tokenValidationContext : new TokenValidationContext(Collections.emptyMap());
+	}
+
+	@Override
+	public void setTokenValidationContext(TokenValidationContext tokenValidationContext) {
+		setRequestAttribute(TOKEN_VALIDATION_CONTEXT_ATTRIBUTE, tokenValidationContext);
+	}
+
+    private Object getRequestAttribute(String name) {
+        return RequestContextHolder.currentRequestAttributes().getAttribute(name, RequestAttributes.SCOPE_REQUEST);
+    }
+
+    private void setRequestAttribute(String name, Object value) {
+        RequestContextHolder.currentRequestAttributes().setAttribute(name, value, RequestAttributes.SCOPE_REQUEST);
+    }
+}
