--- oidc-support/src/main/java/no/nav/security/token/support/core/context/JwtTokenValidationContextHolder.java
@@ -1,12 +1,0 @@
-package no.nav.security.token.support.core.context;
-
-public interface JwtTokenValidationContextHolder {
-
-    Object getRequestAttribute(String name);
-
-    void setRequestAttribute(String name, Object value);
-
-    JwtTokenValidationContext getOIDCValidationContext();
-
-    void setOIDCValidationContext(JwtTokenValidationContext jwtTokenValidationContext);
-}
