+++ oidc-support/src/main/java/no/nav/security/token/support/core/context/TokenContextHolder.java
@@ -1,0 +1,8 @@
+package no.nav.security.token.support.core.context;
+
+public interface TokenContextHolder {
+
+    TokenValidationContext getTokenValidationContext();
+
+    void setTokenValidationContext(TokenValidationContext tokenValidationContext);
+}
