+++ oidc-support/src/main/java/no/nav/security/token/support/core/jwt/JwtTokenClaims.java
@@ -1,0 +1,69 @@
+package no.nav.security.token.support.core.jwt;
+
+import com.nimbusds.jwt.JWTClaimsSet;
+import net.minidev.json.JSONArray;
+
+import java.text.ParseException;
+import java.util.List;
+import java.util.Map;
+
+public class JwtTokenClaims {
+
+    private final JWTClaimsSet jwtClaimsSet;
+
+    public JwtTokenClaims(JWTClaimsSet jwtClaimsSet) {
+        this.jwtClaimsSet = jwtClaimsSet;
+    }
+
+    public Object get(String name) {
+        return getClaimSet().getClaim(name);
+    }
+
+    public String getStringClaim(String name) {
+        try {
+            return getClaimSet().getStringClaim(name);
+        } catch (ParseException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public String getIssuer() {
+        return getClaimSet().getIssuer();
+    }
+
+    public String getSubject() {
+        return getClaimSet().getSubject();
+    }
+
+    public List<String> getAsList(String name) {
+        try {
+            return getClaimSet().getStringListClaim(name);
+        } catch (ParseException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public boolean containsClaim(String name, String value) {
+        Object claim = getClaimSet().getClaim(name);
+        if (claim == null) {
+            return false;
+        }
+        if (claim instanceof String) {
+            String claimAsString = (String) claim;
+            return claimAsString.equals(value);
+        }
+        if (claim instanceof JSONArray) {
+            JSONArray claimasList = (JSONArray) claim;
+            return claimasList.contains(value);
+        }
+        return false;
+    }
+
+    public Map<String, Object> getAllClaims() {
+        return getClaimSet().getClaims();
+    }
+
+    JWTClaimsSet getClaimSet() {
+        return this.jwtClaimsSet;
+    }
+}
