+++ oidc-support/src/main/java/no/nav/security/token/support/core/validation/JwtTokenRetriever.java
@@ -1,7 +1,7 @@
 package no.nav.security.token.support.core.validation;
 
 import no.nav.security.token.support.core.configuration.MultiIssuerConfiguration;
-import no.nav.security.token.support.core.context.JwtToken;
+import no.nav.security.token.support.core.jwt.JwtToken;
 import no.nav.security.token.support.core.http.HttpRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
